requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.8.5
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-appconf==1.0.5
django-ckeditor==6.1.0
django-cors-middleware==1.5.0
django-filter==21.1
django-imagekit==4.1.0
django-js-asset==1.2.2
django-model-utils==4.2.0
django-mptt==0.13.4
django-notifications-hq==1.6.0
django-password-reset==2.0
django-taggit==2.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
flake8==4.0.1
httpie==2.6.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
jedi==0.18.1
Jinja2==3.0.3
jsonfield==3.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.6.0
Markdown==3.3.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nose==1.3.7
notebook==6.4.6
oauthlib==3.1.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.1
QtPy==1.11.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
swapper==1.2.0
sympy==1.8
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yapf==0.31.0
zipp==3.6.0
